Split `ScriptableObjectDatabaseUtility`
Separate `ScriptableObjectDatabaseInspector` into its own file
