+++ Assets/cwg/schwer-scripts-master/schwer-scripts/Database/Editor/ScriptableObjectDatabaseInspector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 87db772437a93b34d9778f86b040c313
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
